refactor: comment dto 패키지를 application 내로 이동
